Jmol.___JmolVersion="14.32.75" // also 15.2.75
bug fix: WRITE ISOSURFACE incorrect for COLOR DENSITY CUTOFF [min,max]

+++ src/org/jmol/jvxl/data/MeshData.java
@@ -350,7 +350,7 @@ public class MeshData extends MeshSurface {
     for (int i = 0; i < n; i++)
       v[i] /= factor;
     if (justOne)
-      return Double.valueOf((double) v[0]);
+      return Double.valueOf(v[0]);
     //System.out.println("MeshData calcVolume " + Escape.e(v));
     if (thisSet != null) {
       thisSet.and(BSUtil.newBitSet2(0, v.length));
